+++ tests/data/test_import_from_illumina.py
@@ -15,7 +15,7 @@ from unittest.mock import patch, PropertyMock
 from plinkio import plinkfile
 
 from src.data.import_from_illumina import main as import_from_illumina
-from src.features.smarterdb import SampleSheep, Dataset, Breed
+from src.features.smarterdb import SampleSheep, Dataset
 
 from ..common import (
     MongoMockMixin, SmarterIDMixin, VariantsMixin, SupportedChipMixin)