+++ tests/common.py
@@ -65,7 +65,8 @@ class SmarterIDMixin():
                 "affyreport.txt",
                 "affyreport_nocols.txt"
             ],
-            type_=["background", "genotypes"]
+            type_=["background", "genotypes"],
+            doi="https://example.com"
         )
         cls.dataset.save()