+++ HISTORY.rst
@@ -19,7 +19,7 @@ TODO
 * Remove ``--chip_name`` from import genotypes script (derive it from dataset)
 * Capitalize ``species_class`` parameter in ``src.data.import_breeds.py``
 
-0.4.7.dev1
+0.4.7.dev2
 ----------
 
 * Import background data from *Gaouar et al 2017* (https://doi.org/10.1038/hdy.2016.86)